Move Backend for feature annotation driver to separate file and add support for file of initial annotations
