+++ varats/varats/tools/driver_feature.py
@@ -1,137 +1,29 @@
 """Driver module for `vara-feature`"""
 import logging
-import re
-import textwrap
 import typing as tp
-from functools import partial, reduce
+from functools import partial
+from pathlib import Path
 
 import click
-import pygit2
-from pygit2 import Walker, Commit, Blob
-from pygit2.enums import SortMode, DiffOption
+from pygit2 import Walker, Commit
+from pygit2.enums import SortMode
 
 from varats.project.project_util import get_local_project_repo
 from varats.tools.tool_util import configuration_lookup_error_handler
-from varats.ts_utils.cli_util import initialize_cli_tool, cli_list_choice
+from varats.ts_utils.cli_util import initialize_cli_tool
 from varats.ts_utils.click_param_types import create_project_choice
-from varats.utils.git_util import CommitHash, FullCommitHash
+from varats.ts_utils.feature_util import (
+    Location,
+    __get_and_check_location,
+    FeatureAnnotation,
+    __get_location_content,
+    __find_potential_new_locations,
+    update_feature_model,
+    load_initial_annotations,
+)
+from varats.utils.git_util import CommitHash
 
 LOG = logging.getLogger(__name__)
-
-
-class Location:
-    """A location in a source code file."""
-
-    LOCATION_FORMAT = re.compile(
-        r"(?P<file>[\w./]+)\s"
-        r"(?P<start_line>\d+):(?P<start_col>\d+)\s"
-        r"(?P<end_line>\d+):(?P<end_col>\d+)"
-    )
-
-    def __init__(
-        self, file: str, start_line: int, start_col: int, end_line: int,
-        end_col: int
-    ) -> None:
-        self.file = file
-        self.start_line = start_line
-        self.start_col = start_col
-        self.end_line = end_line
-        self.end_col = end_col
-
-    @staticmethod
-    def change_start_line(
-        old_location: "Location",
-        new_start_line: int,
-    ):
-        """Move the location the start of the location to a line."""
-        return Location(
-            old_location.file, new_start_line, old_location.start_col,
-            old_location.end_line + (new_start_line - old_location.start_line),
-            old_location.end_col
-        )
-
-    @staticmethod
-    def move_location(old_location: "Location", line_offset: int) -> "Location":
-        """Move the location by a line offset."""
-        return Location.change_start_line(
-            old_location, line_offset + old_location.start_line
-        )
-
-    @staticmethod
-    def parse_string(
-        raw_location: str,
-        old_location: tp.Optional["Location"] = None
-    ) -> "Location":
-        """Create a location from a string."""
-        if old_location and raw_location.isnumeric():
-            new_line = int(raw_location)
-            return Location(
-                old_location.file, new_line, old_location.start_col, new_line,
-                old_location.end_col
-            )
-
-        match = Location.LOCATION_FORMAT.match(raw_location)
-        if match is None:
-            raise click.UsageError(
-                f"Could not parse location: {raw_location}.\n"
-                f"Location format is "
-                f"'<file> <start_line>:<start_col> <end_line>:<end_col>'"
-            )
-
-        return Location(
-            match.group("file"), int(match.group("start_line")),
-            int(match.group("start_col")), int(match.group("end_line")),
-            int(match.group("end_col"))
-        )
-
-    def to_xml(self) -> str:
-        """Convert the location to SPLConqueror feature model format."""
-        xml = f"<path>{self.file}</path>\n"
-        xml += (
-            f"<start><line>{self.start_line}</line>"
-            f"<column>{self.start_col}</column></start>\n"
-        )
-        xml += (
-            f"<end><line>{self.end_line}</line>"
-            f"<column>{self.end_col}</column></end>\n"
-        )
-        return xml
-
-    def __str__(self) -> str:
-        return (
-            f"{self.file} "
-            f"{self.start_line}:{self.start_col} "
-            f"{self.end_line}:{self.end_col}"
-        )
-
-
-class FeatureAnnotation:
-    """A versioned feature source annotation."""
-
-    def __init__(
-        self,
-        feature_name: str,
-        location: Location,
-        introduced: FullCommitHash,
-        removed: tp.Optional[FullCommitHash] = None
-    ) -> None:
-        self.feature_name = feature_name
-        self.location = location
-        self.introduced = introduced
-        self.removed = removed
-
-    def to_xml(self) -> str:
-        """Convert the annotation to SPLConqueror feature model format."""
-        xml = "<sourceRange>\n"
-        xml += "  <revisionRange>\n"
-        xml += f"    <introduced>{self.introduced.hash}</introduced>\n"
-        if self.removed is not None:
-            xml += f"    <removed>{self.removed.hash}</removed>\n"
-        xml += "  </revisionRange>\n"
-        xml += textwrap.indent(self.location.to_xml(), "  ")
-        xml += "</sourceRange>"
-
-        return xml
 
 
 def __prompt_location(
@@ -143,7 +35,6 @@ def __prompt_location(
     default: str = None
 ) -> tp.Tuple[Location, str]:
     commit_hash = CommitHash.from_pygit_commit(commit)
-    LOG.setLevel(logging.DEBUG)
     if prompt is None:
         prompt = f"Enter location for feature {feature_name} @ {commit_hash.short_hash}"
 
@@ -157,102 +48,6 @@ def __prompt_location(
     )
 
 
-def __get_and_check_location(
-    raw_location: str,
-    commit: Commit,
-    old_location: tp.Optional["Location"] = None
-) -> tp.Tuple[Location, str]:
-    location = Location.parse_string(raw_location, old_location)
-    if commit.tree.__contains__(location.file):
-        location_content = __get_location_content(commit, location)
-    else:
-        raise click.UsageError(
-            f"The provided file does not exist in the repository."
-        )
-    if not location_content:
-        raise click.UsageError(
-            f"The provided location does not exist or is empty."
-        )
-
-    return location, location_content
-
-
-def __get_location_content(commit: Commit,
-                           location: Location) -> tp.Optional[str]:
-    lines: tp.List[bytes] = tp.cast(Blob, commit.tree[location.file
-                                                     ]).data.splitlines()
-
-    if len(lines) < location.start_line:
-        return None
-
-    if location.start_line != location.end_line:
-        content = reduce(
-            lambda x, y: x + "\n" + y.decode("utf-8"),
-            lines[location.start_line:location.end_line - 1],
-            lines[location.start_line - 1].decode("utf-8")[location.start_col -
-                                                           1:]
-        )
-        content += "\n" + lines[location.end_line -
-                                1].decode("utf-8")[:location.end_col]
-        return content
-
-    line: str = lines[location.start_line - 1].decode("utf-8")
-    # Todo Why is this end_col and not start_col?
-    if len(line) <= location.end_col:
-        return None
-
-    return line[(location.start_col - 1):location.end_col]
-
-
-def __find_potential_new_locations(
-    repo: pygit2.Repository, commit, annotation: FeatureAnnotation,
-    old_target: str
-) -> tp.List[tp.Tuple[Location, str]]:
-    potential_new_locations: tp.List[tp.Tuple[Location, str]] = []
-    for parent in commit.parents:
-        diff = repo.diff(parent, commit)
-        for patch in diff:
-            if patch.delta.old_file.path == annotation.location.file:
-                offset_counter = 0
-                stop_offset = False
-                for hunk in patch.hunks:
-                    for line in hunk.lines:
-                        if line.old_lineno > annotation.location.end_line:
-                            stop_offset = True
-                        if line.new_lineno >= 0:
-                            if line.old_lineno < 0:
-                                if not stop_offset:
-                                    offset_counter += 1
-                            if annotation.location.end_line == annotation.location.start_line:
-                                content = line.content[
-                                    annotation.location.start_col -
-                                    1:annotation.location.end_col]
-                            else:
-                                content = line.content[annotation.location.
-                                                       start_col - 1:]
-                            if old_target == content:
-                                potential_new_locations.append((
-                                    Location.change_start_line(
-                                        annotation.location, line.new_lineno
-                                    ), content
-                                ))
-                        else:
-                            if not stop_offset:
-                                offset_counter -= 1
-
-                potential_new_location = Location.move_location(
-                    annotation.location, offset_counter
-                )
-                potential_new_content = __get_location_content(
-                    commit, potential_new_location
-                )
-                if potential_new_content == old_target:
-                    potential_new_locations.append(
-                        (potential_new_location, potential_new_content)
-                    )
-    return potential_new_locations
-
-
 @click.group()
 @configuration_lookup_error_handler
 def main() -> None:
@@ -264,6 +59,9 @@ def main() -> None:
 @click.option("--project", "-p", type=create_project_choice(), required=True)
 @click.option("--revision", "-r", type=str, required=False)
 @click.option(
+    "--infile", "-i", type=click.File("r"), default=None, required=False
+)
+@click.option(
     "--outfile",
     "-o",
     type=click.File("w"),
@@ -271,10 +69,10 @@ def main() -> None:
     required=False
 )
 def __annotate(
-    project: str, revision: tp.Optional[str], outfile: tp.TextIO
+    project: str, revision: tp.Optional[str], infile: tp.Optional[tp.TextIO],
+    outfile: tp.TextIO
 ) -> None:
     initialize_cli_tool()
-
     repo = get_local_project_repo(project).pygit_repo
     walker: Walker
 
@@ -289,11 +87,21 @@ def __annotate(
         while first_commit != commit:
             first_commit = next(walker)
 
-    tracked_features: dict[str, dict[int, list[FeatureAnnotation]]] = {}
-    last_annotations: dict[str, dict[int, FeatureAnnotation]] = {}
-    last_annotation_targets: dict[str, dict[int, str]] = {}
+    if infile is not None:
+        LOG.debug("Reading existing annotations from %s", infile.name)
+        tracked_features, last_annotations, last_annotation_targets = load_initial_annotations(
+            infile, first_commit
+        )
+        LOG.debug(
+            f"Loaded {len(tracked_features)} tracked features from {infile.name}."
+        )
+    else:
+        tracked_features: dict[str, dict[int, list[FeatureAnnotation]]] = {}
+        last_annotations: dict[str, dict[int, FeatureAnnotation]] = {}
+        last_annotation_targets: dict[str, dict[int, str]] = {}
 
     LOG.debug("Current revision: %s", first_commit.id)
+
     while click.confirm("Annotate another feature?"):
         feature_name = click.prompt("Enter feature name to annotate", type=str)
         commit_hash = CommitHash.from_pygit_commit(first_commit)
@@ -382,14 +190,17 @@ def __annotate(
             tracked_features[feature][annotation_id].append(annotation)
 
     click.echo(f"Final annotations written to {outfile.name}.")
-    for feature, annotations in tracked_features.items():
-        outfile.write(f"Annotations for feature {feature}:\n")
-        for annotation_id, locations in annotations.items():
-            for location in locations:
-                outfile.write(location.to_xml())
+    if not Path(outfile.name).exists():
+        for feature, annotations in tracked_features.items():
+            outfile.write(f"Annotations for feature {feature}:\n")
+            for annotation_id, locations in annotations.items():
+                for location in locations:
+                    outfile.write(location.to_xml_direct())
+                    outfile.write("\n")
                 outfile.write("\n")
             outfile.write("\n")
-        outfile.write("\n")
+    else:
+        update_feature_model(Path(outfile.name), tracked_features)
 
 
 if __name__ == '__main__':
